API Reference

Public API

WhereTheWaterFlows.waterflows — Function

waterflows(dem, cellarea=fill!(similar(dem),1), flowdir_fn=d8dir_feature;
           feedback_fn=nothing, drain_pits=true, bnd_as_sink=false)

Does the water flow routing according the D8 algorithm. Locations of the dem with NaN-value are ignored.

args:

dem – the DEM (or hydro-potential); array

cellarea=cellarea=fill!(similar(dem),1) – the source per cell, defaults to 1.

if cellarea is negative in places, flux may go to zero but not below.

in areas where no routing takes place, typically NaNs in the dem, cellarea is ignored. This maybe affects mass-conservation.

if using physical units then use a volumetric flux per cell, e.g. m3/s.

Alternatively, cellarea can be a tuple of arrays. Then they are treated/routed separately, for instance (water, tracer). All quantities need to be extensive (i.e. additive, e.g. use internal energy and not temperature)

flowdirfn=d8dirfeature – the routing function. Defaults to the built-in d8dir_feature function but could be customized

kwargs:

feedbackfn – function which is applied to area-value(s) at each cell once all water of the cell has been accumulated but before the water is routed further downstream. Signature `(uparea, ij, dir) -> newuparea--> for example,(uparea, ij, dir) -> max(uparea, 0)` would ensure that all upareas are non-negative.

drain_pits – whether to route through pits (true)

bndassink (true) – whether the domain boundary should be sinks, i.e. adjacent cells can drain into them, or whether to ignore them.

nanassink (true) – whether NaN cells in the DEM should make adjacent cells a sink.

stacksize (2^13 * 2^10) – size of the call-stack in flowroutingcatchments!, which is prone to StackOverflowError. Note however, that OutOfMemory errors are likely if increased.

Returns

area – upslope area (or a tuple of upslope areas if cellarea is a tuple too)

stream-length – length of stream to the farthest source (number of cells traversed)

dir – flow direction at each location

nout – whether the point has outlflow. I.e. nout[I]==0 –> I is a pit

nin – number of inflow cells

sinks – location of sinks as Vector{CartesianIndex{2}}

pits – location of pits as Vector{CartesianIndex{2}}

c – catchment map (color numbers ∈ 1:length(sinks) are for sinks, others for pits)

bnds – boundaries between catchments. The boundary to the exterior/NaNs is not in here.

flowdirextraoutput – extra output of the flowdir_fn, which is nothing for the default

WhereTheWaterFlows.fill_dem — Function

fill_dem(dem, sinks, dir; small=0)

Fill the pits of a DEM (apply this after applying "drainpits", which is done by default in waterflows). Returns the filled DEM.

Notes:

this is not needed as pre-processing step to use the flow-routing function waterflows.

routing on a filled DEM will not produce exactly the same flow pattern: on the shores of lakes streams which entered the lake can now go the other way. I suspect on most DEMs the differences will be very minimal.

This uses a tree traversal to fill the DEM. It does it depth-first (as it is easier) which may lead to a stack overflow on a large DEM.

Post-processing

These functions are exported but are called separately after waterflows:

WhereTheWaterFlows.catchment — Function

catchment(dir, ij)

Calculates the catchment of one or several grid-point(s) ij. If desired, the catchment boundary can be calculated with make_boundaries([c], [1]).

Input

dir – direction field

ij – index of the point (2-Tuple, or CartesianIndex) or a Vector{CartesianIndex} for several

Returns

catchment – BitArray

Tip: only being off by one grid-point can make the difference between a tiny and a huge catchment!

See also: catchments

WhereTheWaterFlows.catchments — Function

catchments(dir, sinks::Union{Vector{Vector{CartesianIndex{2}}}, Vector{<:CartesianIndices{2}}}, dem=nothing;
                check_sinks_overlap=true)

Make a map of catchments from different (non-overlapping) sinks.

See also: catchment

WhereTheWaterFlows.catchment_flux — Function

catchment_flux(cellarea, c, color) = sum(cellarea[c.==color])
catchment_flux(cellarea, c::Union{BitArray, Matrix{Bool}})

The total flux, i.e. input, in one catchment.

WhereTheWaterFlows.prune_catchments — Function

prune_catchments(catchments, minsize; val=0)

Sets all catchments which are smaller than minsize to val (=0). Catchments with number <1 are ignored.

Note that, as the catchments are re-numbered, the number will not correspond to the sinks and pits anymore.

Internals

These functions are not part of the public API but are documented for contributors and users who want to customise the routing pipeline.

WhereTheWaterFlows.d8dir_feature — Function

d8dir_feature(dem, bnd_as_sink, nan_as_sink, extra_sinks=CartesianIndex{2}[], extra_barriers=CartesianIndex{2}[])

D8 directions of a DEM and drainage features (nin & nout).

Elevations with NaN map to dir==BARRIER, cells around them will be set to SINK if nanassink==true or receive no special treatment otherwise.

The argument bnd_as_sink determines whether cells at the domain boundary act as sinks.

The extrasinks and extrabarriers will be added to dir.

Return

dir - direction, encoded as dirnums

nout - number of outflow cells of a cell (0 or 1)

nin - number of inflow cells of a cell (0-8)

sinks - location of sinks as a Vector{CartesianIndex{2}} (sorted) (here dir==SINK)

pits - location of pits as a Vector{CartesianIndex{2}} (sorted) (here dir==PIT)

dem - DEM, unchanged

flowdirextraoutput – nothing (not used by this function, but could be by custom ones)

WhereTheWaterFlows.flowrouting_catchments — Function

flowrouting_catchments(dir, pits, cellarea, feedback_fn, stacksize)

Recursively calculate flow-routing and catchments from

dir - direction field

pits - pit coordinates

cellarea - water input

stacksize - how large the call-stack is (in bytes)

Return:

upslope area

stream length (number of cells traversed)

catchments Matrix{Int}. Value==0 corresponds to NaNs in the DEM which are not pits (i.e. where no water flows into).

Note: this function may cause a stackoverflow on very big catchments.

WhereTheWaterFlows.drainpits! — Function

drainpits!(dir, nin, nout, sinks, pits, ctch, bnds, dem)

Update in place the direction field such that it drains pits. This is done by reversing the flow connecting the lowest point (which can drain) on the catchment boundary to the pit for each such catchment.

Update in place dir, nin, nout, pits (sorted), c; returns an empty bnds

WhereTheWaterFlows.make_boundaries — Function

make_boundaries(catchments, pit_colors, bnds=Origin(firstindex(pit_colors))([CartesianIndex{2}[] for i in 1:length(pit_colors)]) )

Make vectors of boundary cells for catchments of pit_colors (as the name suggests, typically just the pit-catchment colors.) Assumes that pit_colors is sorted.

Note that cells along the edge of the domain[1] as well as cells only bordering BARRIER-cells are not included (because the algorithms do not need to traverse them).

Return:

bnds – Vector of Vector{CartesianIndex{2}} containing the cells which are on the boundary of said catchment.

[1] Note that when bndassink==true then no cells along the boundary will belong to a pit-catchment.

WhereTheWaterFlows.make_flowfeatures — Function

make_flowfeatures(dir)

Makes nout, nin, sinks, pits by iterating once over the dir array

WhereTheWaterFlows.dir2ind — Function

dir2ind(dir, map_special_to_PIT=false)

Translate a D8 direction number into a CartesianIndex (i.e. a flow vector), also maps SINK to CartesianIndex(0,0).

If mapspecialto_PIT==true, then any dir>9 is mapped to CartesianIndex(0,0).

WhereTheWaterFlows.dir2vec — Function

Translate a D8 direction number into a 2D vector.

WhereTheWaterFlows.ind2dir — Function

Translate a CartesianIndex, giving the offset, into a direction number

WhereTheWaterFlows.diagonal_fac — Function

Adjustment factor for gradient in diagonal vs non-diagonal

WhereTheWaterFlows.flowsinto — Function

Tests whether a cell J with flowdir dirJ flows into cell I.

WhereTheWaterFlows.on_outer_boundary — Function

on_outer_boundary(ar, I::CartesianIndex)

Check whether on outer boundary of ar

WhereTheWaterFlows.iterate_D9 — Function

Return CartesianIndices corresponding to the 8 neighbors and the point itself

WhereTheWaterFlows.showme — Function

Show an array with its axes oriented such that they correspond to x and y direction.

WhereTheWaterFlows.dirnums — Constant

Direction numbers. E.g. dirnums[1,1] will return the number corresponding to the direction top-left.

Note, I use the conversion that the x-axis corresponds to the row of the matrix and the y-axis the columns. To print them in this "normal" coordinate system use showme

WhereTheWaterFlows.cartesian — Constant

Translation from dirnums to CartesianIndex

WhereTheWaterFlows._flow_from_to! — Function

Update dir, nin, and nout such that flow at P1 is now from P1 to P2.

It can potentially modify dir, nin, nout at three locations

P1: dir, nout, nin

P2: nin

if allowP2pit==true, then P2's dir, nout can also be modified.

P3 (previous receiver cell of P1): nin

Note that

if P1 and P2 lie in the same catchment, then P3 is also in that catchment.

if flow was from P2 to P1, then P2 has to become a pit (PIT) to keep dir consistent